Changed Doubles to Floats. Changed sensor update method in Skeleton to single arm sensor update. Planning to have different sensor updates for different sensor configurations. This may change back in the future when all sensors are present.

+++ ExerciseMotionTracker/ExerciseMonitor.swift
@@ -18,7 +18,7 @@ public class ExerciseMonitor
         
     }
     
-    public func getPercentComplete() -> Double? {
+    public func getPercentComplete() -> Float? {
         return 0
     }
